requirements.txt
Flask==3.1.2
Python-dotenv
Firebase_Admin==7.1.0
Flask-Mail